fix(routing): deduplicate client_router::history module
The SPA navigation regression chain (#4075 / #4088 / #4122 / #4203 /
#4213 / #4217) recurred six times because reinhardt-pages and
reinhardt-urls kept structurally identical but parallel copies of the
client-side history primitives. Each fix had to be mirrored, and the
mirror was repeatedly missed: PR #4206 fixed pages's copy but left
urls's `JsValue::from_str(&json)` path in place, so downstream
consumers reaching urls via UnifiedRouter still saw the bug.

Move the canonical implementation (post-#4206 fix with structured-JS-
object serialization via `state_to_js_object` / `state_from_js_value`)
to reinhardt-urls and reduce reinhardt-pages's `router::history` to a
thin re-export shim. urls is the lower-level crate (required dep,
versus pages which is optional in the umbrella), so dependency
direction matches the existing `pages -> urls` flow.

Also drop the `pub use` re-exports of `push_state`, `replace_state`,
`current_path`, `go_back`, `go_forward` from
`reinhardt_urls::routers::client_router` itself; verified across the
workspace that no external callers reach them at this path. The
functions remain `pub` inside `history::*` for reinhardt-pages's
cross-crate re-export.

Fixes #4217
