+++ crates/reinhardt-urls/src/routers/client_router.rs
@@ -65,7 +65,11 @@ mod core;
 mod error;
 mod global;
 mod handler;
-mod history;
+// Issue #4217: `history` is exposed publicly so reinhardt-pages can
+// re-export the canonical primitives. The functions inside remain
+// callable cross-crate, but the more ergonomic re-exports at this
+// module level are intentionally limited (see below).
+pub mod history;
 mod params;
 mod pattern;
 mod reverser;
@@ -75,9 +79,11 @@ pub use core::{ClientRoute, ClientRouteMatch, ClientRouter};
 pub use error::{PathError, RouterError};
 pub use global::{clear_client_reverser, get_client_reverser, register_client_reverser};
 pub use handler::RouteHandler;
-pub use history::{
-	HistoryState, NavigationType, current_path, go_back, go_forward, push_state, replace_state,
-};
+// Issue #4217: drop helper-function re-exports from this module's
+// public surface. Callers should use `Router::push()` / `ClientRouter::push()`
+// instead. The functions remain `pub` at `history::*` so reinhardt-pages
+// can re-export them across the crate boundary.
+pub use history::{HistoryState, NavigationType};
 pub use params::{FromPath, ParamContext, Path, SingleFromPath};
 pub use pattern::ClientPathPattern;
 pub use reverser::ClientUrlReverser;
